Refactor server code
- Change `events` to `server_events` to disambiguate with client events
- Move server side state and logic to `App` class
